+++ trtools/mergeSTR/tests/test_mergeSTR.py
@@ -34,9 +34,13 @@ class DummyRecord:
         self.chrom = chrom
         self.pos = pos
         self.info = info
+        self.end_pos = pos + len(ref) - 1
         self.vcfrecord = argparse.Namespace()
         self.vcfrecord.REF = ref
         self.vcfrecord.ALT = alts
+
+    def __str__(self):
+        return self.pos
 
 # Test right files or directory - GangSTR
 def test_GangSTRRightFile(args, mrgvcfdir):
@@ -216,7 +220,9 @@ def test_conflicting_refs_merge(capsys, args, mrgvcfdir):
     for fnames in fname1 + ',' + fname2, fname2 + ',' + fname1:
         args.vcfs = fnames
         main(args)
+
         captured = capsys.readouterr().err
+
         # looking for four skipped records
         # the first has different start coords
         # the second has different end coords
@@ -232,6 +238,75 @@ def test_conflicting_refs_merge(capsys, args, mrgvcfdir):
         assert var.POS == 3045469
         with pytest.raises(StopIteration):
             next(vcf)
+
+def test_conflicting_refs_same_file():
+    recs1 = [DummyRecord('chr1', 100, 'CAG', 'CAGCAG'),
+             DummyRecord('chr1', 100, 'CAG', 'CAGCAGCAG')]
+    recs2 = [DummyRecord('chr1', 100, 'CAG', 'CAGCAG'),
+             DummyRecord('chr1', 1000, 'CAG', 'CAGCAGCAG')]
+    for test_count, readers in enumerate([(iter(recs1), iter(recs2)),
+                                          (iter(recs2), iter(recs1))]):
+        itr = RecordIterator(readers, ['chr1'])
+
+        recs, mins = next(itr)
+        if test_count == 1:
+            # unreverse
+            recs = [recs[1], recs[0]]
+            mins = [mins[1], mins[0]]
+        assert mins == [False, True]
+        assert recs[1].pos == 1000
+        with pytest.raises(StopIteration):
+            next(itr)
+
+
+def test_conflicting_refs_threeway():
+    recs1 = [DummyRecord('chr1', 100, 'CAGCAGCAGCAG'),
+             DummyRecord('chr1', 103, 'CAGCAG'),
+             DummyRecord('chr1', 124, 'CAGCAGCAGCAG'),
+             DummyRecord('chr1', 140, 'CAGCAGCAGCAG')]
+    recs2 = [DummyRecord('chr1', 109, 'CAGCAGCAGCAG'),
+             DummyRecord('chr1', 160, 'CAGCAGCAGCAG'),
+             DummyRecord('chr1', 180, 'CAGCAGCAGCAG')]
+    recs3 = [DummyRecord('chr1', 115, 'CAGCAGCAGCAG'),
+             DummyRecord('chr1', 140, 'CAGCAGCAGCAG'),
+             DummyRecord('chr1', 189, 'CAGCAGCAGCAG')]
+
+    def order(l, idxs):
+        new_l = []
+        new_l.append(l[int(np.where(np.array(idxs) == 0)[0])])
+        new_l.append(l[int(np.where(np.array(idxs) == 1)[0])])
+        new_l.append(l[int(np.where(np.array(idxs) == 2)[0])])
+        return new_l
+
+    list_of_recs = [recs1, recs2, recs3]
+    # repeat this test for any possible ordering of the three files
+    # then reorder the results to be in the same order as displayed
+    # above
+    # only the 140 and 160 records should be returned, everything else
+    # overlaps something else
+    for idxs in [
+        (0,1,2), (0,2,1), (1,0,2), (1,2,0), (2,0,1), (2,1,0)
+    ]:
+        readers = [iter(list_of_recs[idxs[0]]),
+                   iter(list_of_recs[idxs[1]]),
+                   iter(list_of_recs[idxs[2]])]
+        itr = RecordIterator(readers, ['chr1'])
+
+        recs, mins = next(itr)
+        recs = order(recs, idxs)
+        mins = order(mins, idxs)
+        assert mins == [True, False, True]
+        assert recs[0].pos == 140
+        assert recs[2].pos == 140
+
+        recs, mins = next(itr)
+        recs = order(recs, idxs)
+        mins = order(mins, idxs)
+        assert mins == [False, True, False]
+        assert recs[1].pos == 160
+
+        with pytest.raises(StopIteration):
+            next(itr)
 
 
 def test_GetInfoItem(capsys):
@@ -271,6 +346,64 @@ use trtools/testsupport/sample_vcfs/mergeSTR_vcfs/create_test_files.sh
 to regenerate the test files with the new version of mergeSTR.
 """
 
+def _check_merged_output(outputvcffile, inputvcffiles, vcftype):
+    # currently doesn't handle changing chrom
+    # doesn't support changed sample names
+    # don't confirm that the choice of which records
+    # have been merged is correct
+    # Does confirm: for each merged record, confirm that all the data
+    # transferred over is the same as before for each sample
+    outputvcf = None
+    inputvcfs = []
+    try:
+        outputvcf = trh.TRRecordHarmonizer(cyvcf2.VCF(outputvcffile), vcftype)
+        inputvcfs = [
+            trh.TRRecordHarmonizer(cyvcf2.VCF(inputvcffile), vcftype) for
+            inputvcffile in inputvcffiles
+        ]
+        out_samps = outputvcf.vcffile.samples
+        in_samps = [inputvcf.vcffile.samples for inputvcf in inputvcfs]
+        inrecs = [next(inputvcf) for inputvcf in inputvcfs]
+        for rec in outputvcf:
+            for idx in range(len(inputvcfs)):
+                while inrecs[idx] is not None and inrecs[idx].pos < rec.pos:
+                    try:
+                        inrecs[idx] = next(inputvcfs[idx])
+                    except StopIteration:
+                        inrecs[idx] = None
+            for idx, inrec in enumerate(inrecs):
+                if inrec.pos != rec.pos:
+                    continue
+                for info in INFOFIELDS[vcftype]:
+                    if not info[1]:
+                        continue
+                    assert inrec.info[info[0]] == rec.info[info[0]]
+                for old_samp_idx, samp in enumerate(in_samps[idx]):
+                    new_samp_idx = out_samps.index(samp)
+                    assert (rec.GetCalledSamples()[new_samp_idx] ==
+                            inrec.GetCalledSamples()[old_samp_idx])
+                    if not rec.GetCalledSamples()[new_samp_idx]:
+                        continue
+                    assert np.all(inrec.GetStringGenotypes()[old_samp_idx, :] ==
+                                  rec.GetStringGenotypes()[new_samp_idx, :])
+                    for fmt in FORMATFIELDS[vcftype]:
+                        infmt = inrec.format[fmt]
+                        if len(infmt.shape) == 1:
+                            assert np.all(infmt[old_samp_idx] ==
+                                          rec.format[fmt][new_samp_idx])
+                        else:
+                            nans = np.isnan(infmt[old_samp_idx, :])
+                            assert np.all(nans ==
+                                          np.isnan(rec.format[fmt][new_samp_idx, :]))
+                            assert np.all(infmt[old_samp_idx, ~nans] ==
+                                          rec.format[fmt][new_samp_idx, ~nans])
+    finally:
+        if outputvcf is not None:
+            outputvcf.close()
+        for inputvcf in inputvcfs:
+            if inputvcf is not None:
+                inputvcf.close()
+
 
 def test_advntr_output(args, mrgvcfdir):
     fname1 = os.path.join(mrgvcfdir, "test_file_advntr1.vcf.gz")
@@ -279,6 +412,12 @@ def test_advntr_output(args, mrgvcfdir):
     args.vcfs = fname1 + "," + fname2
     assert main(args) == 0
     assert_same_vcf(args.out + '.vcf', mrgvcfdir + "/advntr_merged.vcf")
+    _check_merged_output(
+        args.out + '.vcf',
+        [mrgvcfdir + "/test_file_advntr1.vcf.gz",
+         mrgvcfdir + "/test_file_advntr2.vcf.gz"],
+        trh.VcfTypes.advntr
+    )
 
 
 def test_eh_output(args, mrgvcfdir):
@@ -288,6 +427,12 @@ def test_eh_output(args, mrgvcfdir):
     args.vcfs = fname1 + "," + fname2
     assert main(args) == 0
     assert_same_vcf(args.out + '.vcf', mrgvcfdir + "/eh_merged.vcf")
+    _check_merged_output(
+        args.out + '.vcf',
+        [mrgvcfdir + "/test_file_eh1.vcf.gz",
+         mrgvcfdir + "/test_file_eh2.vcf.gz"],
+        trh.VcfTypes.eh
+    )
 
 
 def test_gangstr_output(args, mrgvcfdir):
@@ -296,6 +441,12 @@ def test_gangstr_output(args, mrgvcfdir):
     args.vcftype = "gangstr"
     args.vcfs = fname1 + "," + fname2
     assert main(args) == 0
+    _check_merged_output(
+        args.out + '.vcf',
+        [mrgvcfdir + "/test_file_gangstr1.vcf.gz",
+         mrgvcfdir + "/test_file_gangstr2.vcf.gz"],
+        trh.VcfTypes.gangstr
+    )
     assert_same_vcf(args.out + '.vcf', mrgvcfdir + "/gangstr_merged.vcf")
 
 
@@ -305,6 +456,12 @@ def test_hipstr_output(args, mrgvcfdir):
     args.vcftype = "hipstr"
     args.vcfs = fname1 + "," + fname2
     assert main(args) == 0
+    _check_merged_output(
+        args.out + '.vcf',
+        [mrgvcfdir + "/test_file_hipstr1.vcf.gz",
+         mrgvcfdir + "/test_file_hipstr2.vcf.gz"],
+        trh.VcfTypes.hipstr
+    )
     assert_same_vcf(args.out + '.vcf', mrgvcfdir + "/hipstr_merged.vcf")
 
 
@@ -314,7 +471,30 @@ def test_popstr_output(args, mrgvcfdir):
     args.vcftype = "popstr"
     args.vcfs = fname1 + "," + fname2
     assert main(args) == 0
+    _check_merged_output(
+        args.out + '.vcf',
+        [mrgvcfdir + "/test_file_popstr1.vcf.gz",
+         mrgvcfdir + "/test_file_popstr2.vcf.gz"],
+        trh.VcfTypes.popstr
+    )
     assert_same_vcf(args.out + '.vcf', mrgvcfdir + "/popstr_merged.vcf")
+
+
+def test_trimmed_hipstr_output(args, mrgvcfdir):
+    fname1 = os.path.join(mrgvcfdir, "NA12878_chr21_hipstr.sorted.vcf.gz")
+    fname2 = os.path.join(mrgvcfdir, "NA12891_chr21_hipstr.sorted.vcf.gz")
+    fname3 = os.path.join(mrgvcfdir, "NA12892_chr21_hipstr.sorted.vcf.gz")
+    args.vcftype = "hipstr"
+    args.trim = True
+    args.vcfs = fname1 + "," + fname2 + "," + fname3
+    assert main(args) == 0
+    _check_merged_output(
+        args.out + '.vcf',
+        [mrgvcfdir + "/NA12878_chr21_hipstr.sorted.vcf.gz",
+         mrgvcfdir + "/NA12891_chr21_hipstr.sorted.vcf.gz",
+         mrgvcfdir + "/NA12892_chr21_hipstr.sorted.vcf.gz"],
+        trh.VcfTypes.hipstr
+    )
 
 
 # TODO questions and issues to confirm, test or  address: